유저 삭제 문제 해결.

+++ src/main/java/com/example/rewrite/service/user/UserServiceImpl.java
@@ -5,8 +5,11 @@ import com.example.rewrite.command.user.ApiResponseDto;
 import com.example.rewrite.command.user.FindIdRequestDto;
 import com.example.rewrite.command.user.SignupRequestDto;
 import com.example.rewrite.command.user.UserDTO;
+import com.example.rewrite.entity.Cart;
 import com.example.rewrite.entity.Product;
 import com.example.rewrite.entity.Users;
+import com.example.rewrite.repository.cart.CartRepository;
+import com.example.rewrite.repository.product.ProductRepository;
 import com.example.rewrite.repository.users.UsersRepository;
 import com.example.rewrite.service.mail.MailService;
 import lombok.RequiredArgsConstructor;
@@ -40,6 +43,8 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     private final UsersRepository usersRepository;
     private final PasswordEncoder passwordEncoder;
     private final MailService mailService;
+    private final ProductRepository productRepository;
+    private final CartRepository cartRepository;
 
     @Override
     @Transactional
@@ -213,7 +218,20 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     @Transactional
     public void deleteUser(Long uid) {
         Users user = usersRepository.findUserByUid(uid);
-        usersRepository.delete(user);
+
+
+        if (user != null) {
+
+            List<Product> products = productRepository.findProductsByUserUid(uid);
+            for (Product product : products) {
+                List<Cart> carts = cartRepository.findCartsByProduct(product);
+                cartRepository.deleteAll(carts);
+            }
+
+            cartRepository.deleteByUserUid(uid); //순서 중요
+            productRepository.deleteByUserUid(uid);
+            usersRepository.delete(user);
+        }
     }
